fixed the binding of flags to show create modals
This branch was created just for the page that shows databases, but the same issue showed in the collections page.

The PR fixes an issue where the “create new database/collection” in the grid view doesn’t actually open the modal to perform the respective action.

The fix was relative simple. The bindings seemed to be done incorrectly, and all I did was fix these bindings.