Update template configuration to show supported runtimes
There are times when people jump into creating the function without
reading the template details. Because templates only support certain
runtime versions, it's important to show the supported runtimes in the
wizard or else the developer will be confused why their enabled
incompatible runtime doesn't show in the Runtime dropdown.