fix: ensure we fetch all projects for usage breakdown
Before this, we tried to fetch all projects in a single query. However,
this fails if there are more than 100 projects because the maximum
number of values that can be included in a query is 100.

Furthermore, the default limit for projects returned by
`projects.list()` is 25.

This PR ensures all projects are fetched by chunking the project IDs
and passing the chunk amount as a limit in the query.

While it is possible to use `projects.get()` instead of
`projects.list()`, the excessive number of network requests ends up
being slower and more overwhelming than using `projects.list()`.